backend/requirements.txt
asgiref==3.9.2
Django==5.2.6
djangorestframework==3.16.1
gunicorn==23.0.0
psycopg2==2.9.10
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.11.0
